#1044 Deleting "all roles" remoes system user and breaks pimcore installation

+++ pimcore/modules/admin/controllers/UserController.php
@@ -212,7 +212,9 @@ class Admin_UserController extends \Pimcore\Controller\Action\Admin
                     $user->delete();
                 }
             } else {
-                $user->delete();
+                if ($user->getId()) {
+                    $user->delete();
+                }
             }
         }
